+++ core/src/main/java/com/metsci/glimpse/event/mouse/newt/MouseWrapperNewt.java
@@ -39,15 +39,15 @@ import com.metsci.glimpse.event.mouse.MouseWrapperImpl;
 public class MouseWrapperNewt extends MouseWrapperImpl<MouseEvent> implements MouseListener
 {
     protected GLWindow glWindow;
-    protected final int scaleX;
-    protected final int scaleY;
+    protected final float scaleX;
+    protected final float scaleY;
 
     public MouseWrapperNewt( NewtGlimpseCanvas canvas )
     {
         super( canvas );
 
         this.glWindow = canvas.getGLWindow( );
-        int[] scale = canvas.getSurfaceScale( );
+        float[] scale = canvas.getSurfaceScale( );
         this.scaleX = scale[0];
         this.scaleY = scale[1];
     }
@@ -55,7 +55,9 @@ public class MouseWrapperNewt extends MouseWrapperImpl<MouseEvent> implements Mo
     @Override
     protected boolean isInterior( MouseEvent e, GlimpseBounds bounds )
     {
-        return bounds.contains( e.getX( ) / scaleX, ( glWindow.getSurfaceHeight( ) - e.getY( ) ) / scaleY );
+        return bounds.contains(
+                (int) ( e.getX( ) / scaleX ),
+                (int) ( ( glWindow.getSurfaceHeight( ) - e.getY( ) ) / scaleY ) );
     }
 
     @Override
@@ -108,7 +110,7 @@ public class MouseWrapperNewt extends MouseWrapperImpl<MouseEvent> implements Mo
 
         for ( int i = 0; i < allX.length; i++ )
         {
-            newX[i] = ( allX[i] / scaleX ) - bounds.getX( );
+            newX[i] = ( (int) ( allX[i] / scaleX ) ) - bounds.getX( );
         }
 
         return newX;
@@ -121,7 +123,9 @@ public class MouseWrapperNewt extends MouseWrapperImpl<MouseEvent> implements Mo
 
         for ( int i = 0; i < allY.length; i++ )
         {
-            newY[i] = e.getY( ) / scaleY - ( parentHeight / scaleY - ( bounds.getY( ) + bounds.getHeight( ) ) );
+            newY[i] = ( (int) ( e.getY( ) / scaleY ) ) -
+                      ( (int) ( parentHeight / scaleY ) ) -
+                      ( bounds.getY( ) + bounds.getHeight( ) );
         }
 
         return newY;
