+++ core/src/main/java/com/metsci/glimpse/gl/util/GLUtils.java
@@ -26,13 +26,13 @@
  */
 package com.metsci.glimpse.gl.util;
 
-import javax.media.opengl.GL;
-import javax.media.opengl.GL3;
-import javax.media.opengl.GLCapabilities;
-import javax.media.opengl.GLContext;
-import javax.media.opengl.GLDrawableFactory;
-import javax.media.opengl.GLOffscreenAutoDrawable;
-import javax.media.opengl.GLProfile;
+import com.jogamp.opengl.GL;
+import com.jogamp.opengl.GL3;
+import com.jogamp.opengl.GLCapabilities;
+import com.jogamp.opengl.GLContext;
+import com.jogamp.opengl.GLDrawableFactory;
+import com.jogamp.opengl.GLOffscreenAutoDrawable;
+import com.jogamp.opengl.GLProfile;
 
 import com.jogamp.opengl.util.FPSAnimator;
 import com.metsci.glimpse.canvas.GlimpseCanvas;
@@ -314,9 +314,9 @@ public class GLUtils
 
     public static void setViewportAndScissor( GlimpseContext context )
     {
-        final int[] scale = context.getSurfaceScale( );
-        final int scaleX = scale[0];
-        final int scaleY = scale[1];
+        final float[] scale = context.getSurfaceScale( );
+        final float scaleX = scale[0];
+        final float scaleY = scale[1];
         GL gl = context.getGL( );
 
         GlimpseBounds bounds = context.getTargetStack( ).getBounds( );
@@ -324,8 +324,17 @@ public class GLUtils
 
         gl.glEnable( GL.GL_SCISSOR_TEST );
 
-        gl.glViewport( bounds.getX( ) * scaleX, bounds.getY( ) * scaleY, bounds.getWidth( ) * scaleX, bounds.getHeight( ) * scaleY );
-        gl.glScissor( clippedBounds.getX( ) * scaleX, clippedBounds.getY( ) * scaleY, clippedBounds.getWidth( ) * scaleX, clippedBounds.getHeight( ) * scaleY );
+        gl.glViewport(
+                (int) ( bounds.getX( ) * scaleX ),
+                (int) ( bounds.getY( ) * scaleY ),
+                (int) ( bounds.getWidth( ) * scaleX ),
+                (int) ( bounds.getHeight( ) * scaleY ) );
+        
+        gl.glScissor(
+                (int) ( clippedBounds.getX( ) * scaleX ),
+                (int) ( clippedBounds.getY( ) * scaleY ),
+                (int) ( clippedBounds.getWidth( ) * scaleX ),
+                (int) ( clippedBounds.getHeight( ) * scaleY ) );
     }
 
     public static GlimpseBounds getClippedBounds( GlimpseContext context )
